Move SchemaCommandRequest to submodule
